back_env/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
google-auth
openai
PyJWT
pymupdf
python-dotenv
requests
drf_yasg
